+++ apps/backend/common/png_logger.py
