+++ apps/backend/state_mgmt_layer/data_per_driver/car_info.py
@@ -27,7 +27,7 @@ from typing import Optional
 
 from lib.f1_types import CarStatusData
 from lib.fuel_rate_recommender import FuelRateRecommender
-from apps.backend.png_logger import getLogger
+from apps.backend.common.png_logger import getLogger
 
 # -------------------------------------- GLOBALS -----------------------------------------------------------------------
