+++ apps/backend/state_mgmt_layer/data_per_driver/packet_copies.py
@@ -30,7 +30,7 @@ from lib.f1_types import (CarDamageData, CarMotionData, CarSetupData,
                           FinalClassificationData, LapData,
                           PacketSessionHistoryData, PacketTyreSetsData,
                           ParticipantData)
-from apps.backend.png_logger import getLogger
+from apps.backend.common.png_logger import getLogger
 
 # -------------------------------------- GLOBALS -----------------------------------------------------------------------
