+++ apps/backend/state_mgmt_layer/data_per_driver/tyre_info.py
@@ -29,7 +29,7 @@ from typing import Any, Dict, List, Optional
 from lib.f1_types import (ActualTyreCompound, PacketTyreSetsData,
                           VisualTyreCompound)
 from lib.tyre_wear_extrapolator import TyreWearExtrapolator, TyreWearPerLap
-from ..png_logger import getLogger
+from apps.backend.common.png_logger import getLogger
 
 # -------------------------------------- GLOBALS -----------------------------------------------------------------------
