+++ apps/backend/state_mgmt_layer/telemetry_web_api.py
@@ -25,13 +25,13 @@
 from typing import Any, Dict, List, Optional, Union
 
 import lib.race_analyzer as RaceAnalyzer
-import apps.backend.telemetry_data as TelData
+import apps.backend.state_mgmt_layer.telemetry_data as TelData
 from lib.f1_types import (CarStatusData, F1Utils, LapHistoryData,
                           VisualTyreCompound)
 from lib.tyre_wear_extrapolator import TyreWearPerLap
-from apps.backend.data_per_driver import DataPerDriver, TyreSetInfo
-from apps.backend.overtakes import GetOvertakesStatus
-from apps.backend.png_logger import getLogger
+from apps.backend.state_mgmt_layer.data_per_driver import DataPerDriver, TyreSetInfo
+from apps.backend.state_mgmt_layer.overtakes import GetOvertakesStatus
+from apps.backend.common.png_logger import getLogger
 
 # -------------------------------------- GLOBALS -----------------------------------------------------------------------
