+++ apps/backend/telemetry_layer/telemetry_handler.py
@@ -31,14 +31,14 @@ from datetime import datetime
 from typing import Any, Dict, List, Optional, Tuple
 
 import lib.race_analyzer as RaceAnalyzer
-import apps.backend.telemetry_data as TelData
+import apps.backend.state_mgmt_layer.telemetry_data as TelData
 from lib.button_debouncer import ButtonDebouncer
 from lib.f1_types import (F1PacketType, PacketEventData,
                           PacketFinalClassificationData, PacketSessionData,
                           SessionType23, SessionType24)
 from lib.inter_task_communicator import AsyncInterTaskCommunicator
 from lib.packet_forwarder import AsyncUDPForwarder
-from apps.backend.png_logger import getLogger
+from apps.backend.common.png_logger import getLogger
 from lib.telemetry_manager import AsyncF1TelemetryManager
 
 # -------------------------------------- TYPE DEFINITIONS --------------------------------------------------------------
